requirements.txt
anthropic==0.49.0
beautifulsoup4==4.13.4
datasets==3.5.0
fastapi==0.115.12
huggingface_hub==0.30.2
jinja2==3.1.3
keyring==25.6.0
keyrings.alt==5.0.0
markdown==3.6
python-multipart==0.0.9
websockets==11.0.3
langchain_anthropic==0.3.11
langchain_core==0.3.52
langchain_experimental==0.3.4
langchain_neo4j==0.4.0
langchain_openai==0.3.13
openai==1.75.0
openai-agents==0.0.11
playwright==1.51.0
pydantic==2.11.3
PyGithub==2.6.1
PyPDF2==3.0.1
pytest==8.3.5
python-dotenv==1.1.0
python_crontab==3.2.0
Requests==2.32.3
torch==2.6.0
transformers==4.51.3
uvicorn==0.34.1
